// models/Content.ts
export interface Content {
  id: number;
  homeHeader: string;
  homeContent: string;
  about: string;
  aboutMission: string;
  aboutVision: string;
  contactPhone: string;
  contactPhoneSecond: string;
  contactAddress: string;
  contactMail: string;
  contactMailSecond: string;
  contactFax: string;
}